requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs
opencensus
opencensus-ext-azure
botbuilder-ai
botbuilder-applicationinsights
botbuilder-azure
botbuilder-schema
botbuilder-core
botbuilder-integration-aiohttp>=4.14.0
botbuilder-integration-applicationinsights-aiohttp
botbuilder-dialogs
botframework-connector
aiohttp
msrest==0.6.10
datatypes-date-time>=1.0.0.a1